+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/steps/medlemskap/MedlemskapStep.tsx
@@ -10,7 +10,6 @@ import getIntlFormErrorHandler from '@navikt/sif-common-formik-ds/src/validation
 import { ValidationError } from '@navikt/sif-common-formik-ds/src/validation/types';
 import BostedUtlandListAndDialog from '@navikt/sif-common-forms-ds/src/forms/bosted-utland/BostedUtlandListAndDialog';
 import { Utenlandsopphold } from '@navikt/sif-common-forms-ds/src/forms/utenlandsopphold/types';
-import { dateToday } from '@navikt/sif-common-utils/src/dateUtils';
 import PersistStepFormValues from '../../../components/persist-step-form-values/PersistStepFormValues';
 import { useOnValidSubmit } from '../../../hooks/useOnValidSubmit';
 import { useStepNavigation } from '../../../hooks/useStepNavigation';
@@ -30,6 +29,7 @@ import {
     getMedlemskapStepInitialValues,
     getMedlemskapSøknadsdataFromFormValues,
 } from './medlemskapStepUtils';
+import { getDateToday } from '@navikt/sif-common-utils';
 
 export enum MedlemskapFormFields {
     harBoddUtenforNorgeSiste12Mnd = 'harBoddUtenforNorgeSiste12Mnd',
@@ -53,7 +53,7 @@ const { FormikWrapper, Form, YesOrNoQuestion } = getTypedFormComponents<
 
 const MedlemskapStep = () => {
     const { text, intl } = useAppIntl();
-    const { neste12Måneder, siste12Måneder } = getMedlemskapDateRanges(dateToday);
+    const { neste12Måneder, siste12Måneder } = getMedlemskapDateRanges(getDateToday());
     const {
         state: { søknadsdata },
     } = useSøknadContext();