+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/steps/oppsummering/components/MedlemskapOppsummering.tsx
@@ -3,7 +3,7 @@ import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
 import { SummarySection } from '@navikt/sif-common-ui';
 import { SummaryList } from '@navikt/sif-common-ui';
 import { JaNeiSvar, SummaryBlock } from '@navikt/sif-common-ui';
-import { dateToday, ISODateToDate } from '@navikt/sif-common-utils/src';
+import { ISODateToDate, getDateToday } from '@navikt/sif-common-utils/src';
 import dayjs from 'dayjs';
 import isSameOrAfter from 'dayjs/plugin/isSameOrAfter';
 import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
@@ -21,8 +21,8 @@ export interface Props {
 const MedlemskapOppsummering: React.FC<Props> = ({ bosteder }) => {
     const { text } = useAppIntl();
 
-    const bostederSiste12 = bosteder.filter((b) => dayjs(ISODateToDate(b.tilOgMed)).isSameOrBefore(dateToday));
-    const bostederNeste12 = bosteder.filter((b) => dayjs(ISODateToDate(b.tilOgMed)).isSameOrAfter(dateToday));
+    const bostederSiste12 = bosteder.filter((b) => dayjs(ISODateToDate(b.tilOgMed)).isSameOrBefore(getDateToday()));
+    const bostederNeste12 = bosteder.filter((b) => dayjs(ISODateToDate(b.tilOgMed)).isSameOrAfter(getDateToday()));
 
     return (
         <SummarySection header={text('step.oppsummering.medlemskap.header')}>