+++ apps/omsorgspengerutbetaling-soknad/src/app/søknad/steps/medlemskap/medlemskapFieldValidations.ts
@@ -1,5 +1,5 @@
 import { Utenlandsopphold } from '@navikt/sif-common-forms-ds';
-import { date1YearAgo, date1YearFromNow } from '@navikt/sif-common-utils';
+import { getDate1YearAgo, getDate1YearFromNow } from '@navikt/sif-common-utils';
 import { dateRangesCollide, dateRangesExceedsRange } from '@navikt/sif-common-utils';
 import dayjs from 'dayjs';
 
@@ -17,7 +17,7 @@ export const validateUtenlandsoppholdSiste12Mnd = (utenlandsopphold: Utenlandsop
     if (dateRangesCollide(dateRanges)) {
         return UtenlandsoppholdErrors.utenlandsopphold_overlapper;
     }
-    if (dateRangesExceedsRange(dateRanges, { from: date1YearAgo, to: dayjs().subtract(1, 'day').toDate() })) {
+    if (dateRangesExceedsRange(dateRanges, { from: getDate1YearAgo(), to: dayjs().subtract(1, 'day').toDate() })) {
         return UtenlandsoppholdErrors.utenlandsopphold_utenfor_periode;
     }
     return undefined;
@@ -31,7 +31,7 @@ export const validateUtenlandsoppholdNeste12Mnd = (utenlandsopphold: Utenlandsop
     if (dateRangesCollide(dateRanges)) {
         return UtenlandsoppholdErrors.utenlandsopphold_overlapper;
     }
-    if (dateRangesExceedsRange(dateRanges, { from: new Date(), to: date1YearFromNow })) {
+    if (dateRangesExceedsRange(dateRanges, { from: new Date(), to: getDate1YearFromNow() })) {
         return UtenlandsoppholdErrors.utenlandsopphold_utenfor_periode;
     }
     return undefined;