+++ apps/omsorgspengerutbetaling-soknad/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @navikt/omsorgspengerutbetaling-soknad
 
+## 1.11.9
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+-   Updated dependencies
+    -   @navikt/sif-common-formik-ds@1.8.7
+    -   @navikt/sif-common-soknad-ds@10.15.7
+    -   @navikt/sif-common-forms-ds@8.22.2
+    -   @navikt/sif-common-core-ds@9.17.8
+    -   @navikt/sif-common-utils@3.46.2
+
+## 1.11.8
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+-   Updated dependencies
+    -   @navikt/sif-common-amplitude@2.25.5
+    -   @navikt/sif-common-formik-ds@1.8.6
+    -   @navikt/sif-common-soknad-ds@10.15.6
+    -   @navikt/sif-common-forms-ds@8.22.1
+    -   @navikt/sif-common-core-ds@9.17.7
+    -   @navikt/sif-common-sentry@0.26.5
+    -   @navikt/sif-common-utils@3.46.1
+
+## 1.11.7
+
+### Patch Changes
+
+-   Updated dependencies [ecd2cdb]
+    -   @navikt/sif-common-forms-ds@8.22.0
+    -   @navikt/sif-common-utils@3.46.0
+    -   @navikt/sif-common-core-ds@9.17.6
+    -   @navikt/sif-common-soknad-ds@10.15.5
+
+## 1.11.6
+
+### Patch Changes
+
+-   6751f58: Pakkeoppdateringer - minor/patch
+-   Updated dependencies [6751f58]
+    -   @navikt/sif-common-amplitude@2.25.4
+    -   @navikt/sif-common-formik-ds@1.8.5
+    -   @navikt/sif-common-soknad-ds@10.15.5
+    -   @navikt/sif-common-forms-ds@8.21.6
+    -   @navikt/sif-common-core-ds@9.17.5
+    -   @navikt/sif-common-sentry@0.26.4
+    -   @navikt/sif-common-utils@3.45.4
+
 ## 1.11.5
 
 ### Patch Changes