+++ packages/sif-common-core-ds/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @navikt/sif-common-core-ds
 
+## 9.17.8
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+-   Updated dependencies
+    -   @navikt/sif-common-utils@3.46.2
+
+## 9.17.7
+
+### Patch Changes
+
+-   Pakkeoppdateringer - minor/patch
+-   Updated dependencies
+    -   @navikt/sif-common-utils@3.46.1
+
+## 9.17.6
+
+### Patch Changes
+
+-   Updated dependencies [ecd2cdb]
+    -   @navikt/sif-common-utils@3.46.0
+
+## 9.17.5
+
+### Patch Changes
+
+-   6751f58: Pakkeoppdateringer - minor/patch
+-   Updated dependencies [6751f58]
+    -   @navikt/sif-common-utils@3.45.4
+
 ## 9.17.4
 
 ### Patch Changes